requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.20
pydantic==1.10.7
python-dotenv==1.0.0
psycopg2-binary==2.9.6